+++ src/main/java/sample/Game/Mechanics/MainMechanics.java
@@ -22,36 +22,6 @@ import java.util.*;
  * Created by ksg on 25.04.17.
  */
 
-
-@SuppressWarnings("PublicField")
-class GameContext {
-    public final Map<String, GameUserItem> mp;
-    public final Map<String, GameCard> table;
-    public final GameCard[] cards;
-    public final Queue<GameUserItem> masterQeue;
-    public CardDeck deck;
-    public GameState state;
-    public int numberOfPlayers;
-    public ObjectMapper mapper;
-    public int numberCardsInHand;
-    public int currentRound;
-    public GameUserItem master;
-
-    public void reset() {
-        table.clear();
-        Arrays.fill(cards, null);
-    }
-
-    public GameContext() {
-        state = new InitState(this);
-        mp = new HashMap<>(); // TODO: Do it as user controller
-        table = new LinkedHashMap<>();
-        masterQeue = new ArrayDeque<>();
-        cards = new GameCard[3];
-        currentRound = 0;
-    }
-}
-
 @SuppressWarnings("DefaultFileTemplate")
 public class MainMechanics {
     public enum ErrorCodes {
